Added documentation. Started sample sepration for ease of digestion Started work on using pooling when deseralizing the message arrays.

+++ src/NexNetDemo/Samples/SampleBase.cs
@@ -1,7 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
+using System.Net.Security;
 using System.Net.Sockets;
+using System.Security.Authentication;
+using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
 using NexNet.Transports;
@@ -10,24 +14,74 @@ namespace NexNetDemo.Samples;
 
 public class SampleBase
 {
-    protected UdsServerConfig ServerConfig { get; }
-    protected UdsClientConfig ClientConfig { get; }
-
-    public SampleBase(bool log)
+    public enum TransportMode
     {
-        var path = "test.sock";
-        if (File.Exists(path))
-            File.Delete(path);
+        Uds,
+        Tcp,
+        TlsTcp
+    }
+    protected ServerConfig ServerConfig { get; }
+    protected ClientConfig ClientConfig { get; }
 
-        ServerConfig = new UdsServerConfig()
+    public SampleBase(bool log, TransportMode transportMode)
+    {
+        if (transportMode == TransportMode.Uds)
         {
-            EndPoint = new UnixDomainSocketEndPoint(path),
-            Logger = log ? new Logger("Server") : null
-        };
-        ClientConfig = new UdsClientConfig()
+            var path = "test.sock";
+            if (File.Exists(path))
+                File.Delete(path);
+
+            ServerConfig = new UdsServerConfig()
+            {
+                EndPoint = new UnixDomainSocketEndPoint(path), 
+                Logger = log ? new Logger("Server") : null
+            };
+            ClientConfig = new UdsClientConfig()
+            {
+                EndPoint = new UnixDomainSocketEndPoint(path),
+                Logger = log ? new Logger("Client") : null
+            };
+        }
+        else if (transportMode == TransportMode.Tcp)
         {
-            EndPoint = new UnixDomainSocketEndPoint(path),
-            Logger = log ? new Logger("Client") : null
-        };
+            ServerConfig = new TcpServerConfig()
+            {
+                EndPoint = new IPEndPoint(IPAddress.Loopback, 1236),
+                Logger = log ? new Logger("Server") : null
+            };
+            ClientConfig = new TcpClientConfig()
+            {
+                EndPoint = new IPEndPoint(IPAddress.Loopback, 1236),
+                Logger = log ? new Logger("Client") : null
+            };
+        }
+        else if (transportMode == TransportMode.TlsTcp)
+        {
+            ServerConfig = new TcpTlsServerConfig()
+            {
+                EndPoint = new IPEndPoint(IPAddress.Loopback, 1236),
+                Logger = log ? new Logger("Server") : null,
+                SslServerAuthenticationOptions = new SslServerAuthenticationOptions()
+                {
+                    CertificateRevocationCheckMode = X509RevocationMode.NoCheck,
+                    ClientCertificateRequired = false,
+                    AllowRenegotiation = false,
+                    EnabledSslProtocols = SslProtocols.Tls12 | SslProtocols.Tls13,
+                    ServerCertificate = new X509Certificate2("server.pfx", "certPass"),
+                },
+            };
+            ClientConfig = new TcpTlsClientConfig()
+            {
+                EndPoint = new IPEndPoint(IPAddress.Loopback, 1236),
+                Logger = log ? new Logger("Client") : null,
+                SslClientAuthenticationOptions = new SslClientAuthenticationOptions()
+                {
+                    EnabledSslProtocols = SslProtocols.Tls12 | SslProtocols.Tls13,
+                    CertificateRevocationCheckMode = X509RevocationMode.NoCheck,
+                    AllowRenegotiation = false,
+                    RemoteCertificateValidationCallback = (sender, certificate, chain, errors) => true
+                }
+            };
+        }
     }
 }
